--- integration-kinesis-to-lambda-sample-code/Function.cs
@@ -1,45 +1,0 @@
-﻿using System.Text;
-using Amazon.Lambda.Core;
-using Amazon.Lambda.KinesisEvents;
-
-// Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
-[assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
-
-namespace KinesisIntegrationSampleCode;
-
-public class Function
-{
-    public async Task FunctionHandler(KinesisEvent evnt, ILambdaContext context)
-    {
-        if (evnt.Records.Count == 0)
-        {
-            context.Logger.LogLine("Empty Kinesis Event received");
-            return;
-        }
-
-        foreach (var record in evnt.Records)
-        {
-            try
-            {
-                context.Logger.LogLine($"Processed Event with EventId: {record.EventId}");
-                string data = await GetRecordDataAsync(record.Kinesis, context);
-                context.Logger.LogLine($"Data: {data}");
-                // TODO: Do interesting work based on the new data
-            }
-            catch (Exception ex)
-            {
-                context.Logger.LogError($"An error occurred {ex.Message}");
-                throw;
-            }
-        }
-        context.Logger.LogLine($"Successfully processed {evnt.Records.Count} records.");
-    }
-
-    private async Task<string> GetRecordDataAsync(KinesisEvent.Record record, ILambdaContext context)
-    {
-        byte[] bytes = record.Data.ToArray();
-        string data = Encoding.UTF8.GetString(bytes);
-        await Task.CompletedTask; //Placeholder for actual async work
-        return data;
-    }
-}
