+++ integration-kinesis-to-lambda/Function.cs
@@ -1,39 +1,20 @@
 ﻿using System.Text;
-using System.Text.Json.Serialization;
 using Amazon.Lambda.Core;
 using Amazon.Lambda.KinesisEvents;
 
 // Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
-//[assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
+[assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
 
-namespace KinesisIntegration;
+namespace KinesisIntegrationSampleCode;
 
 public class Function
 {
-    /// <summary>
-    /// Default constructor. This constructor is used by Lambda to construct the instance. When invoked in a Lambda environment
-    /// the AWS credentials will come from the IAM role associated with the function and the AWS region will be set to the
-    /// region the Lambda function is executed in.
-    /// </summary>
-    public Function()
-    {
-
-    }
-
-
-    /// <summary>
-    /// This method is called for every Lambda invocation. This method takes in an Kinesis event object and can be used 
-    /// to respond to Kinesis stream records.
-    /// </summary>
-    /// <param name="evnt"></param>
-    /// <param name="context"></param>
-    /// <returns></returns>
-    public async Task<StreamsEventResponse> FunctionHandler(KinesisEvent evnt, ILambdaContext context)
+    public async Task FunctionHandler(KinesisEvent evnt, ILambdaContext context)
     {
         if (evnt.Records.Count == 0)
         {
             context.Logger.LogLine("Empty Kinesis Event received");
-            return new StreamsEventResponse();
+            return;
         }
 
         foreach (var record in evnt.Records)
@@ -48,19 +29,10 @@ public class Function
             catch (Exception ex)
             {
                 context.Logger.LogError($"An error occurred {ex.Message}");
-                /* Since we are working with streams, we can return the failed item immediately.
-                   Lambda will immediately begin to retry processing from this failed item onwards. */
-                return new StreamsEventResponse
-                {
-                    BatchItemFailures = new List<StreamsEventResponse.BatchItemFailure>
-                    {
-                        new StreamsEventResponse.BatchItemFailure { ItemIdentifier = record.Kinesis.SequenceNumber }
-                    }
-                };
+                throw;
             }
         }
         context.Logger.LogLine($"Successfully processed {evnt.Records.Count} records.");
-        return new StreamsEventResponse();
     }
 
     private async Task<string> GetRecordDataAsync(KinesisEvent.Record record, ILambdaContext context)
@@ -69,16 +41,5 @@ public class Function
         string data = Encoding.UTF8.GetString(bytes);
         await Task.CompletedTask; //Placeholder for actual async work
         return data;
-    }
-}
-
-public class StreamsEventResponse
-{
-    [JsonPropertyName("batchItemFailures")]
-    public IList<BatchItemFailure> BatchItemFailures { get; set; }
-    public class BatchItemFailure
-    {
-        [JsonPropertyName("itemIdentifier")]
-        public string ItemIdentifier { get; set; }
     }
 }
